Rename: 로또 만들고 계산하는 클래스를 파일로 묶음
- lotto-make-and-calculate 파일을 만들었다.
- 그리고 로또 만들기, 랭킹 계산하기, 수익률 계산하기 클래스를 넣었다.
