+++ docs/README.md
@@ -1,11 +1,16 @@
 ## 기능 목록
-- [x] 로또 구입 가격을 1000원 단위로 입력받는다. - UserInput#getPrice()
+- [x] 로또 구입 가격을 1000원 단위로 입력받는다.
+    ```
+    <예외상황>
+    1000단위로 나누어 떨어지지 않는 경우
+    ```
 
-- [x] 가격에 따른 수량만큼 로또 발행(1~45 중 중복 없는 6개의 수) - makeLotto()
+- [x] 가격에 따른 수량만큼 로또 발행(1~45 중 중복 없는 6개의 수)
 
 - [x] 수량과 발행된 로또 출력(오름차순)
 
 - [x] (,)를 기준으로 당첨 번호 받기
+    - [x] 유효값 검사
     ```
     <예외상황>
     - 6개 초과,미만을 받음
@@ -14,12 +19,13 @@
     - (,)이외의 문자가 있는 경우
     ```
 - [x] 당첨번호와 일치하지 않는 보너스 번호 입력 받기
-    '''
+    - [x] 유효값 검사
+    ```
     <예외상황>
     - 숫자가 아닐 경우
     - 당첨번호에 있는 수와 중복될 경우
     - 범위를 넘어서는 경우
-    '''
+    ```
 
 - [x] 각 로또마다 몇 개의 수가 맞았는지 계산한다.
     - [x] 당첨 통계 출력
@@ -31,6 +37,3 @@
 
 - [x] 수익률 계산기능
     - [x] 수익률 출력 (소수점 둘째 자리에서 반올림)
-
-
-- 사용자가 잘못된 값을 입력할 경우 `throw`문을 사용해 예외를 발생시키고, "[ERROR]"로 시작하는 에러 메시지를 출력 후 종료한다.
