+++ src/lotto-make-and-calculate/Calculate.js
